saas/requirements.txt
Django>=5.1,<5.2
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui==1.5.3
django-widget-tweaks
slippers
stripe
django-compressor
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
django-crispy-forms
crispy-bootstrap5